Исправляет нейминг классов
