+++ src/Service/Actions/CancelAction.php
@@ -4,7 +4,7 @@ namespace Taskforce\Service\Actions;
 
 use Taskforce\Logic\Task;
 
-class ActionCancel extends Action
+class ActionCancel extends BaseAction
 {
 
     protected string $actionName = 'Отменить задание';
